Add package.json
This project's goal is to be simple and direct to the point. This
should be the answer to the order: "I need an Web Application and I
need it now!"

I consider it a Development Workflow, one of the goals are:
* be easier to start working here, than start working on a CopePen.

* Be easy to deploy to GitHub Pages or to deploy to my own Digital
Ocean droplet.

* It should increase my creative time while working. It means: less
time wondering why my workflow isn't working and more time to wonder
how awesome is my Web Application.

As said in 4d42cbe this list will increase as this project get improved.