Add npm start
npm start will start a local server into the folder `src/`, it also
will open the browser with the -o

I noticed the line 9 was changed, but would not it be better if we
doesn't had that comma? In the diff would be included only the new
lines. I took I look and there's is repositories that uses a
package.yaml to allow comments and remove the brackets.
https://www.npmjs.com/package/npm-yaml

I think YAML is awesome, would be awesome if npm accept both JSON and
YAML formats.