Add deploy to GitHub Pages
GitHub Pages are awesome, because you just need to create a branch
called `gh-pages` that all the files in the branch will be served into
a custom URL to your project on github.com.

If your Github user is `github.com/githuber/` and your repository is
`github.com/githuber/repository`, if you access
`githuber.github.io/repository` you can see the `index.html` of the
branch `ng-pages`.

That's how GitHub Page works, and it's freaking simple, you can use to
host all your projects and event setup a custom domain, we'll probably
cover custom domain in the future.

With reference in the great gist
https://gist.github.com/cobyism/4730490 I added the command `npm run
gh-pages`.

When you run this, it will get the content of the `src` folder and send
to `gh-pages` branch, you can check the result here
https://github.com/felquis/workflow-web/tree/gh-pages

Again we cant have a clear diff, because the comma mentioned in 66cb986
it added complexity to the diff and I still thinking package.yaml is an
awesome solution.