Ignore *.out.js
I decided to use an ES6 transpiler, I'm looking for an easy and
intuitive way to work with JavaScript. And ES6 modules are awesome,
Arrow Functions are awesome and a lot of ohter ES6 features are awesome.

I believe that from 2015 to the future, we will always transpile our
codes, it help us to stay in touch with the most recent features of
ES6, 7... and it's always a good idea stay in touch with the in-coming
features.

The transpiled files will be saved with the `name.out.js`, with the
.out.js suffix.

Let's see what's coming in the next commit :)