Add sublime-project files
As Sublime Text is my default IDE, I think it's natural to try to use
the maximum this could can provide me. So, I decided to keep a
configuration file, this config will change the aspect of the file side
bar inside Sublime, it is good to hide auto-generated files from the
side bar, to focus just in the code.

:) Let's continue.